Arsenal_V4/webpanel-backup/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
requests==2.31.0
PyJWT==2.8.0
python-dateutil==2.8.2
psutil==5.9.5
